requirements.txt
pygame==2.6.0
pyyaml
speechmatics-python
sounddevice
setuptools